Add unit test framework and 2 module tests.
This commit adds the test framework, and unit tests for console and fs.
To facilitate the repeated construction and destruction of modules,
the utility functions have been moved into module-utils.ks and its
functionality has expanded to include the following changes:

A register function can be called to register the module and a check for
whether the module is properly loaded

The require function now checks with the map of registered modules.
If it exists in the map, then it checks if the module is properly loaded.
If not properly loaded, it force reloads the module (using runPath
instead of runOncePath). If the module is loaded properly, nothing
occurs. if it doesn't exist in the map, then runOncePath is invoked.

A utility function has been added to validate that all registered modules
are properly loaded, and if they aren't force load them.

Note: if a module does *not* invoke the register method, then it will
only be subject to the runOncePath mechanism, so if a mandatory part of
the module gets unset, then the module system will not be able to force
reload the module. A script would have to re-run the module manually.